@@ -1,7 +1,8 @@
 import { BrowserModule, Title } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+// import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { HighchartsChartModule } from 'highcharts-angular';
 import { Shared } from '../shared/shared.module';
 import { routing } from '../app.routes';
@@ -30,7 +31,7 @@ import { ClipboardService } from '../clipboard.service';
     HighchartsChartModule,
     FormsModule,
     ReactiveFormsModule,
-    HttpModule
+    HttpClientModule
   ],
   entryComponents: [
     CohHelpComponent
