@@ -1,8 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterLinkStubDirective, ActivatedRouteStub, ActivatedRoute } from '../../testing/router-stubs';
 import { RouterTestingModule } from '@angular/router/testing';
-import { HttpModule, Http, XHRBackend, BaseRequestOptions, ConnectionBackend, Response, ResponseOptions } from '@angular/http';
-
+import { XHRBackend, BaseRequestOptions, ConnectionBackend, Response, ResponseOptions } from '@angular/http';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { ShareLinkComponent } from './share-link.component';
 import { ClipboardService } from '../clipboard.service';
 import { AnalyzerService } from '../analyzer.service';
@@ -36,7 +36,7 @@ describe('ShareLinkComponent', () => {
         { provide: 'rootCategory', useValue: 59 },
         { provide: ActivatedRoute, useValue: activatedRoute }
       ],
-      imports: [ HttpModule, RouterTestingModule ]
+      imports: [ HttpClientModule, RouterTestingModule ]
     })
     .compileComponents();
   }));
