@@ -15,7 +15,7 @@ import 'core-js/es6/map';
 import 'core-js/es6/set';
 import 'core-js/es6/reflect';
 
-import 'core-js/es7/reflect';
+
 import 'zone.js/dist/zone';
 
 // If you need to support the browsers/features below, uncomment the import
